fix(telescope) Workaround lsp_dynamic_workspace_symbols
Nvim v0.11 has a small bug where using goto reference/ goto definition
can error. This is fixed in the git branch of telescope, but not yet in
stable. This should be changed back once this is brought into stable
telescope. This might be a while though, considering the most recent
version of telescope is from may 2024 (almost a year ago).